mHMMbayes: multilevel hidden Markov models using Bayesian estimation.

With the R package mHMMbayes you can fit multilevel hidden Markov models. The multilevel hidden Markov model (HMM) is a generalization of the well-known hidden Markov model, tailored to accommodate (intense) longitudinal data of multiple individuals simultaneously. Using a multilevel framework, we allow for heterogeneity in the model parameters (transition probability matrix and conditional distribution), while estimating one overall HMM. The model has a great potential of application in many fields, such as the social sciences and medicine. The model can be fitted on multivariate data with a categorical distribution, and include individual level covariates (allowing for e.g., group comparisons on model parameters). Parameters are estimated using Bayesian estimation utilizing the forward-backward recursion within a hybrid Metropolis within Gibbs sampler.

Details

The mHMMbayes package provides three main functions: mHMM , sim_mHMM and vit_mHMM, described below. For a more elaborate guide to the package mHMMbayes, see the tutorial-mhmm vignette: vignette("tutorial-mhmm", package = "mHMMbayes") . For extensive information on the estimation of the parameters in the package, see the estimation-mhmm vignette: vignette("estimation-mhmm", package = "mHMMbayes").

mHMM

The function mHMM fits a multilevel hidden Markov model to (intense longitudinal) data from multiple subjects using Bayesian estimation. By using a multilevel framework, one general 'population' HMM is estimated, while heterogeneity between subjects is accommodated. The function can handle covariates at the subject level varying observation lengths over subjects. Estimation is performed using a hybrid Metropolis within Gibbs sampler, and completes the forward backward algorithm for all subjects in a sequential manner.

sim_mHMM

The function sim_mHMM simulates data for multiple subjects, for which the data have categorical observations that follow a hidden Markov model (HMM) with an multilevel structure. The multilevel structure implies that each subject is allowed to have it's own set of parameters, and that the parameters at the subject level (level 1) are tied together by a population distribution at level 2 for each of the corresponding parameters. The shape of the population distribution for each of the parameters is a normal (i.e., Gaussian) distribution. In addition to (natural and/or unexplained) heterogeneity between subjects, the subjects parameters can also depend on a (set of) covariate(s).

vit_mHMM

The function vit_mHMM obtains the most likely hidden state sequence for each subject, given the data and the subject specific parameter estimates. The function does this by utilizing the Viterbi algorithm.