Obtain the emission distribution probabilities for a fitted multilevel HMM

obtain_emiss obtains the emission distribution for a fitted multilevel hidden Markov model, for either the group level, i.e., representing the average emission distribution over all subjects, or at the subject level, returning the emission distribution for each subject.

Usage

obtain_emiss(object, level = "group", burn_in = NULL)

Arguments

object

An object of class mHMM, generated by the function mHMM.

level

String specifying if the returned transition probability matrix gamma should be at the group level (level = "group"), i.e., representing the average transition probability matrix over all subjects, or at the subject level (level = "subject").

burn_in

An integer which specifies the number of iterations to discard when obtaining the model parameter summary statistics. When left unspecified (burn_in = NULL), the burn in period specified when creating the mHMM object will be used.

Value

obtain_emiss returns the object est_emiss. Depending on the specification at the input variable level, est_emiss is either a list of matrices with the emission distribution at the group level (if level = "group") for each dependent variable, or a list of lists, where for each dependent variable a list is returned with the number of elements equal to the number of subjects analyzed, if level = 'subject'). In the latter scenario, each matrix in the lower level list represents the subject specific emission distribution for a specific dependent variable.

See also

Examples

###### Example on package data, see ?nonverbal
# \donttest{
# specifying general model properties:
m <- 2
n_dep <- 4
q_emiss <- c(3, 2, 3, 2)

# specifying starting values
start_TM <- diag(.8, m)
start_TM[lower.tri(start_TM) | upper.tri(start_TM)] <- .2
start_EM <- list(matrix(c(0.05, 0.90, 0.05,
                          0.90, 0.05, 0.05), byrow = TRUE,
                        nrow = m, ncol = q_emiss[1]), # vocalizing patient
                 matrix(c(0.1, 0.9,
                          0.1, 0.9), byrow = TRUE, nrow = m,
                        ncol = q_emiss[2]), # looking patient
                 matrix(c(0.90, 0.05, 0.05,
                          0.05, 0.90, 0.05), byrow = TRUE,
                        nrow = m, ncol = q_emiss[3]), # vocalizing therapist
                 matrix(c(0.1, 0.9,
                          0.1, 0.9), byrow = TRUE, nrow = m,
                        ncol = q_emiss[4])) # looking therapist

# Run a model without covariate(s):
out_2st <- mHMM(s_data = nonverbal,
                gen = list(m = m, n_dep = n_dep, q_emiss = q_emiss),
                start_val = c(list(start_TM), start_EM),
                mcmc = list(J = 11, burn_in = 5))
#> Progress of the Bayesian mHMM algorithm: 
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%  |                                                                              |================                                                      |  22%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================                                       |  44%  |                                                                              |=======================================                               |  56%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%  |                                                                              |==============================================================        |  89%  |                                                                              |======================================================================| 100%
#> Total time elapsed (hh:mm:ss): 00:00:05

out_2st
#> Number of subjects: 10 
#> 
#> 11 iterations used in the MCMC algorithm with a burn in of 5 
#> Average Log likelihood over all subjects: -1637.137 
#> Average AIC over all subjects: 3302.273 
#> 
#> Number of states used: 2 
#> 
#> Number of dependent variables used: 4 
#> 
summary(out_2st)
#> State transition probability matrix 
#>  (at the group level): 
#>  
#>              To state 1 To state 2
#> From state 1      0.934      0.066
#> From state 2      0.072      0.928
#> 
#>  
#> Emission distribution for each of the dependent variables 
#>  (at the group level): 
#>  
#> $p_vocalizing
#>         Category 1 Category 2 Category 3
#> State 1      0.030      0.930       0.04
#> State 2      0.757      0.103       0.14
#> 
#> $p_looking
#>         Category 1 Category 2
#> State 1      0.261      0.739
#> State 2      0.166      0.834
#> 
#> $t_vocalizing
#>         Category 1 Category 2 Category 3
#> State 1      0.801      0.100      0.098
#> State 2      0.034      0.942      0.024
#> 
#> $t_looking
#>         Category 1 Category 2
#> State 1      0.039      0.961
#> State 2      0.275      0.725
#> 
#> 

# obtaining the emission probabilities at the group and subject level
obtain_emiss(out_2st, level = "group")
#> $p_vocalizing
#>         Category 1 Category 2 Category 3
#> State 1      0.030      0.930       0.04
#> State 2      0.757      0.103       0.14
#> 
#> $p_looking
#>         Category 1 Category 2
#> State 1      0.261      0.739
#> State 2      0.166      0.834
#> 
#> $t_vocalizing
#>         Category 1 Category 2 Category 3
#> State 1      0.801      0.100      0.098
#> State 2      0.034      0.942      0.024
#> 
#> $t_looking
#>         Category 1 Category 2
#> State 1      0.039      0.961
#> State 2      0.275      0.725
#> 
obtain_emiss(out_2st, level = "subject")
#> $p_vocalizing
#> $p_vocalizing$`Subject 1`
#>         Category 1 Category 2 Category 3
#> State 1      0.008      0.987      0.005
#> State 2      0.843      0.045      0.113
#> 
#> $p_vocalizing$`Subject 2`
#>         Category 1 Category 2 Category 3
#> State 1      0.028      0.827      0.146
#> State 2      0.849      0.048      0.103
#> 
#> $p_vocalizing$`Subject 3`
#>         Category 1 Category 2 Category 3
#> State 1      0.043      0.938      0.020
#> State 2      0.836      0.045      0.119
#> 
#> $p_vocalizing$`Subject 4`
#>         Category 1 Category 2 Category 3
#> State 1      0.019      0.954      0.027
#> State 2      0.831      0.068      0.102
#> 
#> $p_vocalizing$`Subject 5`
#>         Category 1 Category 2 Category 3
#> State 1      0.002      0.989      0.009
#> State 2      0.754      0.168      0.077
#> 
#> $p_vocalizing$`Subject 6`
#>         Category 1 Category 2 Category 3
#> State 1      0.017      0.951      0.032
#> State 2      0.694      0.108      0.198
#> 
#> $p_vocalizing$`Subject 7`
#>         Category 1 Category 2 Category 3
#> State 1      0.162      0.784      0.054
#> State 2      0.827      0.048      0.125
#> 
#> $p_vocalizing$`Subject 8`
#>         Category 1 Category 2 Category 3
#> State 1      0.012      0.963      0.026
#> State 2      0.436      0.408      0.155
#> 
#> $p_vocalizing$`Subject 9`
#>         Category 1 Category 2 Category 3
#> State 1      0.102      0.837      0.061
#> State 2      0.747      0.129      0.124
#> 
#> $p_vocalizing$`Subject 10`
#>         Category 1 Category 2 Category 3
#> State 1      0.009      0.966      0.025
#> State 2      0.783      0.107      0.110
#> 
#> 
#> $p_looking
#> $p_looking$`Subject 1`
#>         Category 1 Category 2
#> State 1      0.121      0.879
#> State 2      0.037      0.963
#> 
#> $p_looking$`Subject 2`
#>         Category 1 Category 2
#> State 1      0.300      0.700
#> State 2      0.068      0.932
#> 
#> $p_looking$`Subject 3`
#>         Category 1 Category 2
#> State 1      0.455      0.545
#> State 2      0.181      0.819
#> 
#> $p_looking$`Subject 4`
#>         Category 1 Category 2
#> State 1      0.069      0.931
#> State 2      0.015      0.986
#> 
#> $p_looking$`Subject 5`
#>         Category 1 Category 2
#> State 1      0.220      0.780
#> State 2      0.104      0.896
#> 
#> $p_looking$`Subject 6`
#>         Category 1 Category 2
#> State 1      0.081      0.919
#> State 2      0.016      0.984
#> 
#> $p_looking$`Subject 7`
#>         Category 1 Category 2
#> State 1      0.444      0.556
#> State 2      0.210      0.790
#> 
#> $p_looking$`Subject 8`
#>         Category 1 Category 2
#> State 1      0.355      0.645
#> State 2      0.183      0.817
#> 
#> $p_looking$`Subject 9`
#>         Category 1 Category 2
#> State 1      0.233      0.767
#> State 2      0.102      0.898
#> 
#> $p_looking$`Subject 10`
#>         Category 1 Category 2
#> State 1      0.446      0.554
#> State 2      0.168      0.832
#> 
#> 
#> $t_vocalizing
#> $t_vocalizing$`Subject 1`
#>         Category 1 Category 2 Category 3
#> State 1      0.822      0.124      0.054
#> State 2      0.020      0.973      0.007
#> 
#> $t_vocalizing$`Subject 2`
#>         Category 1 Category 2 Category 3
#> State 1      0.822      0.134      0.044
#> State 2      0.060      0.874      0.066
#> 
#> $t_vocalizing$`Subject 3`
#>         Category 1 Category 2 Category 3
#> State 1      0.855      0.043      0.102
#> State 2      0.035      0.927      0.038
#> 
#> $t_vocalizing$`Subject 4`
#>         Category 1 Category 2 Category 3
#> State 1      0.755      0.107      0.138
#> State 2      0.014      0.977      0.010
#> 
#> $t_vocalizing$`Subject 5`
#>         Category 1 Category 2 Category 3
#> State 1      0.814      0.083      0.104
#> State 2      0.012      0.970      0.018
#> 
#> $t_vocalizing$`Subject 6`
#>         Category 1 Category 2 Category 3
#> State 1      0.808      0.113      0.078
#> State 2      0.023      0.926      0.051
#> 
#> $t_vocalizing$`Subject 7`
#>         Category 1 Category 2 Category 3
#> State 1      0.736      0.192      0.072
#> State 2      0.038      0.957      0.005
#> 
#> $t_vocalizing$`Subject 8`
#>         Category 1 Category 2 Category 3
#> State 1      0.629      0.088      0.283
#> State 2      0.026      0.959      0.014
#> 
#> $t_vocalizing$`Subject 9`
#>         Category 1 Category 2 Category 3
#> State 1       0.75      0.059      0.191
#> State 2       0.03      0.964      0.006
#> 
#> $t_vocalizing$`Subject 10`
#>         Category 1 Category 2 Category 3
#> State 1      0.813      0.103      0.084
#> State 2      0.008      0.972      0.020
#> 
#> 
#> $t_looking
#> $t_looking$`Subject 1`
#>         Category 1 Category 2
#> State 1      0.016      0.984
#> State 2      0.194      0.806
#> 
#> $t_looking$`Subject 2`
#>         Category 1 Category 2
#> State 1      0.037      0.963
#> State 2      0.295      0.705
#> 
#> $t_looking$`Subject 3`
#>         Category 1 Category 2
#> State 1      0.010      0.990
#> State 2      0.246      0.754
#> 
#> $t_looking$`Subject 4`
#>         Category 1 Category 2
#> State 1      0.015      0.985
#> State 2      0.261      0.739
#> 
#> $t_looking$`Subject 5`
#>         Category 1 Category 2
#> State 1      0.011      0.989
#> State 2      0.202      0.797
#> 
#> $t_looking$`Subject 6`
#>         Category 1 Category 2
#> State 1      0.017      0.983
#> State 2      0.304      0.696
#> 
#> $t_looking$`Subject 7`
#>         Category 1 Category 2
#> State 1      0.092      0.908
#> State 2      0.218      0.782
#> 
#> $t_looking$`Subject 8`
#>         Category 1 Category 2
#> State 1      0.184      0.816
#> State 2      0.191      0.809
#> 
#> $t_looking$`Subject 9`
#>         Category 1 Category 2
#> State 1      0.336      0.664
#> State 2      0.230      0.770
#> 
#> $t_looking$`Subject 10`
#>         Category 1 Category 2
#> State 1      0.041      0.959
#> State 2      0.422      0.578
#> 
#> 

# }