Obtain the transition probabilities gamma for a fitted multilevel HMM

obtain_gamma obtains the transition probability matrix for a fitted multilevel hidden Markov model, for either the group level, i.e., representing the average transition probability matrix over all subjects, or at the subject level, returning the transition probability matrices for each subject.

Usage

obtain_gamma(object, level = "group", burn_in = NULL)

Arguments

object

An object of class mHMM, generated by the function mHMM.

level

String specifying if the returned transition probability matrix gamma should be at the group level (level = "group"), i.e., representing the average transition probability matrix over all subjects, or at the subject level (level = "subject").

burn_in

An integer which specifies the number of iterations to discard when obtaining the model parameter summary statistics. When left unspecified (burn_in = NULL), the burn in period specified when creating the mHMM object will be used.

Value

obtain_gamma returns the object est_gamma of the class

mHMM_gamma. This object can be directly plotted using the function

plot.mHMM_gamma(), or simply plot(). Depending on the specification at the input variable level, est_gamma is either a matrix with the transition probabilities at the group level (if

level = "group"), or a list of matrices (with the number of elements equal to the number of subjects analyzed, if level = 'subject'), where each matrix in the list represents a subject specific transition probability matrix.

See also

Examples

###### Example on package data, see ?nonverbal
# \donttest{
# specifying general model properties:
m <- 2
n_dep <- 4
q_emiss <- c(3, 2, 3, 2)

# specifying starting values
start_TM <- diag(.8, m)
start_TM[lower.tri(start_TM) | upper.tri(start_TM)] <- .2
start_EM <- list(matrix(c(0.05, 0.90, 0.05,
                          0.90, 0.05, 0.05), byrow = TRUE,
                        nrow = m, ncol = q_emiss[1]), # vocalizing patient
                 matrix(c(0.1, 0.9,
                          0.1, 0.9), byrow = TRUE, nrow = m,
                        ncol = q_emiss[2]), # looking patient
                 matrix(c(0.90, 0.05, 0.05,
                          0.05, 0.90, 0.05), byrow = TRUE,
                        nrow = m, ncol = q_emiss[3]), # vocalizing therapist
                 matrix(c(0.1, 0.9,
                          0.1, 0.9), byrow = TRUE, nrow = m,
                        ncol = q_emiss[4])) # looking therapist

# Run a model without covariate(s):
out_2st <- mHMM(s_data = nonverbal,
                gen = list(m = m, n_dep = n_dep, q_emiss = q_emiss),
                start_val = c(list(start_TM), start_EM),
                mcmc = list(J = 11, burn_in = 5))
#> Progress of the Bayesian mHMM algorithm: 
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%  |                                                                              |================                                                      |  22%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================                                       |  44%  |                                                                              |=======================================                               |  56%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%  |                                                                              |==============================================================        |  89%  |                                                                              |======================================================================| 100%
#> Total time elapsed (hh:mm:ss): 00:00:05

out_2st
#> Number of subjects: 10 
#> 
#> 11 iterations used in the MCMC algorithm with a burn in of 5 
#> Average Log likelihood over all subjects: -1640.904 
#> Average AIC over all subjects: 3309.808 
#> 
#> Number of states used: 2 
#> 
#> Number of dependent variables used: 4 
#> 
summary(out_2st)
#> State transition probability matrix 
#>  (at the group level): 
#>  
#>              To state 1 To state 2
#> From state 1      0.933      0.067
#> From state 2      0.073      0.927
#> 
#>  
#> Emission distribution for each of the dependent variables 
#>  (at the group level): 
#>  
#> $p_vocalizing
#>         Category 1 Category 2 Category 3
#> State 1      0.025      0.952      0.023
#> State 2      0.753      0.088      0.159
#> 
#> $p_looking
#>         Category 1 Category 2
#> State 1      0.261      0.739
#> State 2      0.112      0.888
#> 
#> $t_vocalizing
#>         Category 1 Category 2 Category 3
#> State 1      0.784      0.109      0.107
#> State 2      0.051      0.913      0.036
#> 
#> $t_looking
#>         Category 1 Category 2
#> State 1      0.049      0.951
#> State 2      0.292      0.708
#> 
#> 

# obtaining the transition probabilities at the group and subject level
obtain_gamma(out_2st, level = "group")
#>              To state 1 To state 2
#> From state 1      0.933      0.067
#> From state 2      0.073      0.927
obtain_gamma(out_2st, level = "subject")
#> $`Subject 1`
#>              To state 1 To state 2
#> From state 1      0.956      0.044
#> From state 2      0.051      0.949
#> 
#> $`Subject 2`
#>              To state 1 To state 2
#> From state 1      0.937      0.063
#> From state 2      0.050      0.950
#> 
#> $`Subject 3`
#>              To state 1 To state 2
#> From state 1      0.961      0.039
#> From state 2      0.045      0.955
#> 
#> $`Subject 4`
#>              To state 1 To state 2
#> From state 1      0.938      0.062
#> From state 2      0.034      0.966
#> 
#> $`Subject 5`
#>              To state 1 To state 2
#> From state 1      0.953      0.047
#> From state 2      0.057      0.943
#> 
#> $`Subject 6`
#>              To state 1 To state 2
#> From state 1      0.949      0.052
#> From state 2      0.095      0.905
#> 
#> $`Subject 7`
#>              To state 1 To state 2
#> From state 1      0.925      0.075
#> From state 2      0.050      0.950
#> 
#> $`Subject 8`
#>              To state 1 To state 2
#> From state 1      0.957      0.043
#> From state 2      0.059      0.941
#> 
#> $`Subject 9`
#>              To state 1 To state 2
#> From state 1      0.961      0.039
#> From state 2      0.046      0.954
#> 
#> $`Subject 10`
#>              To state 1 To state 2
#> From state 1      0.968      0.032
#> From state 2      0.060      0.940
#> 

# }